fix: package nome e inserção das viewmodel
