--- APP/app/src/main/java/view/navigation/maingest/MainGestViewModel.kt
@@ -1,2 +1,0 @@
-package view.navigation.maingest
-
