--- APP/app/src/main/java/view/navigation/singup/SingnupViewModel.kt
@@ -1,2 +1,0 @@
-package view.navigation.singup
-
