+++ APP/app/src/main/java/view/pages/maingest/MainGest.kt
@@ -1,15 +1,12 @@
-package view.navigation.maingest
+package view.pages.maingest
 
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import navigation.Routes
-import view.navigation.login.LoginForm
 
 @Composable
 fun MainGest() {
@@ -20,7 +17,7 @@ fun MainGest() {
             startDestination = Routes.maingest,
             modifier = Modifier.padding(innerPadding)
         ) {
-            
+
         }
     }
 }
