+++ APP/app/src/main/java/view/pages/singup/SingnupViewModel.kt
@@ -1,0 +1,2 @@
+package view.pages.singup
+
